+++ HISTORY.md
@@ -8,6 +8,8 @@
   Thanks @Retsam.
 - Fixed: #493: function `combinations` did not throw an exception for 
   non-integer values of `k`.
+- Fixed: function `import` did not override typed functions when the option
+  override was set true.
 
 
 ## 2015-10-09, version 2.4.0